explicitly include stdint.h to pdm_can.c

+++ PDM/Inc/can/pdm_can.h
@@ -4,5 +4,5 @@
 #endif //PDM_PDM_CAN_H
 void pdm_can_init();
 
-void pdm_can_tx(const uint8_t tx_data[], uint8_t tx_data_len, uint32_t tx_id);
-void pdm_can_rx(uint8_t rx_data[], uint8_t rx_data_len, uint32_t rx_id);
+void pdm_can_tx(const uint8_t tx_data[], uint8_t tx_data_len, uint16_t tx_id);
+void pdm_can_rx(uint8_t rx_data[], uint8_t rx_data_len, uint16_t rx_id);
